// src/lib/utils/canonicalTzRegions.generated.ts
// Generated by scripts/generate-globe-data.ts
// Source: Natural Earth 10m Time Zones

export const CANONICAL_TZ_REGIONS = [
  '+1',
  '+11',
  '+2',
  '+3',
  '+5',
  '+5.5',
  '+6',
  '+7',
  '+9',
  '-10',
  '-11',
  '-12',
  '-2',
  '-3',
  '-4',
  '-8',
  '-9',
  '0',
  'Africa/Johannesburg',
  'America/Anchorage',
  'America/Caracas',
  'America/Chicago',
  'America/Danmarkshavn',
  'America/Denver',
  'America/La_Paz',
  'America/Los_Angeles',
  'America/Miquelon',
  'America/New_York',
  'America/Sao_Paulo',
  'America/Scoresbysund',
  'America/St_Johns',
  'Antarctica/',
  'Antarctica/Casey',
  'Antarctica/Central',
  'Antarctica/DumontDUrville',
  'Antarctica/Mawson',
  'Antarctica/Mirny',
  'Antarctica/Palmer',
  'Antarctica/South_Pole',
  'Antarctica/Troll',
  'Asia/Almaty',
  'Asia/Dhaka',
  'Asia/Dubai',
  'Asia/Irkutsk',
  'Asia/Jakarta',
  'Asia/Kabul',
  'Asia/Kamchatka',
  'Asia/Karachi',
  'Asia/Kathmandu',
  'Asia/Kolkata',
  'Asia/Omsk',
  'Asia/Rangoon',
  'Asia/Riyadh',
  'Asia/Tashkent',
  'Asia/Tehran',
  'Asia/Thimphu',
  'Asia/Tokyo',
  'Asia/Vladivostok',
  'Asia/Yakutsk',
  'Atlantic/Cape_Verde',
  'Atlantic/South_Georgia',
  'Australia/Adelaide',
  'Australia/Eucla',
  'Australia/Lord_Howe',
  'Australia/Perth',
  'Australia/Sydney',
  'Europe/Helsinki',
  'Europe/Kaliningrad',
  'Europe/Lisbon',
  'Europe/London',
  'Europe/Mariehamn',
  'Europe/Minsk',
  'Europe/Moscow',
  'Europe/Paris',
  'Indian/Chagos',
  'Indian/Cocos',
  'Pacific/Auckland',
  'Pacific/Chatham',
  'Pacific/Enderbury',
  'Pacific/Guadalcanal',
  'Pacific/Honolulu',
  'Pacific/Kiritimati',
  'Pacific/Marquesas',
  'Pacific/Midway',
  'Pacific/Norfolk',
  'Pacific/Pago_Pago',
] as const

export type CanonicalTzRegion = (typeof CANONICAL_TZ_REGIONS)[number]